fix(memory): cool down retries after incomplete search maintenance
A detached search maintenance can resolve while its manager stays
dirty (for example a keyword-only provider fallback): the generation
is restored to the serving manager and an incomplete reason is
reported instead of throwing. The retry gate treated that as a
success and reset its cooldown, so the next search could immediately
relaunch a full rebuild. Count an incomplete outcome as a failure for
the escalating cooldown, and reset the escalation only after an
attempt fully clears the handed-off generation.

Move the retry-gate regressions into a dedicated test file and add
coverage driving the public search() path, including an end-to-end
case where the real detached maintenance generation publishes new
content without any maintenance-manager stub.

Addresses review feedback for #135754
