+++ extensions/memory-core/src/memory/manager-search-orchestration.test.ts
@@ -899,101 +899,4 @@ describe("memory index", () => {
     await expect(manager.close?.()).resolves.toBeUndefined();
     expect(syncSpy).toHaveBeenCalledWith({ reason: "search" });
   });
-
-  it("coalesces concurrent search-triggered maintenance into a single detached attempt", async () => {
-    const manager = await getPersistentManager(
-      createCfg({ provider: "none", minScore: 0, onSearch: true, hybrid: { enabled: true } }),
-    );
-    await manager.sync({ reason: "test" });
-    const maintenanceStarted = createDeferred<void>();
-    const releaseMaintenance = createDeferred<void>();
-    const maintenance = {
-      sync: vi.fn(async () => {
-        maintenanceStarted.resolve();
-        await releaseMaintenance.promise;
-      }),
-      status: vi.fn(() => ({ dirty: false })),
-      close: vi.fn(async () => {}),
-    };
-    const getSpy = vi.spyOn(MemoryIndexManager, "get").mockResolvedValue(maintenance as never);
-    Reflect.set(manager, "dirty", true);
-    const fields = manager as unknown as {
-      syncPublishedIndexInBackground: (params: { reason: string }) => Promise<void>;
-    };
-
-    try {
-      const first = fields.syncPublishedIndexInBackground({ reason: "search" });
-      await maintenanceStarted.promise;
-      const second = fields.syncPublishedIndexInBackground({ reason: "search" });
-      // Give a wrongly-queued second attempt a chance to acquire another manager.
-      await new Promise((resolve) => {
-        setTimeout(resolve, 50);
-      });
-      expect(getSpy).toHaveBeenCalledTimes(1);
-      expect(maintenance.sync).toHaveBeenCalledTimes(1);
-
-      releaseMaintenance.resolve();
-      await Promise.allSettled([first, second]);
-      expect(getSpy).toHaveBeenCalledTimes(1);
-      expect(maintenance.close).toHaveBeenCalledTimes(1);
-    } finally {
-      releaseMaintenance.resolve();
-      getSpy.mockRestore();
-      await manager.close?.();
-    }
-  });
-
-  it("defers search-triggered maintenance retries after a failed attempt and recovers later", async () => {
-    const manager = await getPersistentManager(
-      createCfg({ provider: "none", minScore: 0, onSearch: true, hybrid: { enabled: true } }),
-    );
-    await manager.sync({ reason: "test" });
-    const syncError = new Error("maintenance failed");
-    let maintenanceAttempts = 0;
-    const maintenance = {
-      sync: vi.fn(async () => {
-        maintenanceAttempts += 1;
-        if (maintenanceAttempts === 1) {
-          throw syncError;
-        }
-      }),
-      status: vi.fn(() => ({ dirty: false })),
-      close: vi.fn(async () => {}),
-    };
-    const getSpy = vi.spyOn(MemoryIndexManager, "get").mockResolvedValue(maintenance as never);
-    Reflect.set(manager, "dirty", true);
-    Reflect.set(manager, "memoryFullRetryDirty", true);
-    const fields = manager as unknown as {
-      syncPublishedIndexInBackground: (params: { reason: string }) => Promise<void>;
-    };
-
-    try {
-      await expect(fields.syncPublishedIndexInBackground({ reason: "search" })).rejects.toThrow(
-        syncError,
-      );
-      expect(maintenance.sync).toHaveBeenCalledTimes(1);
-      expect(Reflect.get(manager, "dirty")).toBe(true);
-      expect(Reflect.get(manager, "memoryFullRetryDirty")).toBe(true);
-
-      // An immediate retry stays deferred: no new detached maintenance starts.
-      await fields.syncPublishedIndexInBackground({ reason: "search" });
-      expect(maintenance.sync).toHaveBeenCalledTimes(1);
-
-      // After the failure cooldown elapses, a later search retries and recovers.
-      vi.useFakeTimers({ toFake: ["Date"] });
-      try {
-        vi.setSystemTime(new Date(Date.now() + 30 * 60_000));
-        await fields.syncPublishedIndexInBackground({ reason: "search" });
-        expect(maintenance.sync).toHaveBeenCalledTimes(2);
-        expect(Reflect.get(manager, "dirty")).toBe(false);
-        expect(Reflect.get(manager, "memoryFullRetryDirty")).toBe(false);
-        expect(manager.status().lastSyncError).toBeUndefined();
-      } finally {
-        vi.useRealTimers();
-      }
-    } finally {
-      getSpy.mockRestore();
-      await manager.close?.();
-    }
-  });
 });
